Download

Function ProcessDownloadTask

pub(crate) async fn ProcessDownloadTask(
    Task: DownloadTask,
    Client: Client,
    Cache: Arc<Mutex<DownloadCache>>,
) -> Result<()>

The main asynchronous function for processing a single download task. This function is designed to be run concurrently for multiple tasks.